Replace `handlers::commands` to `handlers.rs`
