+++ justfile
@@ -20,7 +20,7 @@ compose-run:
 compose-run-build:
     docker compose up --build
 
-# This is just my template to run databse without `just compose-run`. If you want to override it, then use it:
+# This is just my template to run databse without `just compose-run`. If you want to override it, then use this template:
 # docker run --rm --name {NAME} -p {PORT}:5432 -e POSTGRES_PASSWORD={PASSWORD} -e POSTGRES_USER={USER} -e POSTGRES_DB={DATABASE_NAME} postgres:17-alpine
 run-db:
     docker run --rm --name steal_stickers_bot_db \
